chore: cleanup tuils module, move utc_now_iso to common.py
