+++ packages/devqubit-braket/src/devqubit_braket/adapter.py
@@ -38,8 +38,8 @@ from devqubit_braket.tracked import TrackedTask, TrackedTaskBatch
 from devqubit_braket.utils import extract_task_id, get_backend_name
 from devqubit_engine.core.run import Run
 from devqubit_engine.uec.envelope import ExecutionEnvelope
+from devqubit_engine.utils.common import utc_now_iso
 from devqubit_engine.utils.serialization import to_jsonable
-from devqubit_engine.utils.time_utils import utc_now_iso
 
 
 logger = logging.getLogger(__name__)
