+++ packages/devqubit-braket/src/devqubit_braket/envelope.py
@@ -52,8 +52,8 @@ from devqubit_engine.uec.types import (
     ProgramRole,
     TranspilationMode,
 )
+from devqubit_engine.utils.common import utc_now_iso
 from devqubit_engine.utils.serialization import to_jsonable
-from devqubit_engine.utils.time_utils import utc_now_iso
 
 
 if TYPE_CHECKING:
