+++ packages/devqubit-braket/src/devqubit_braket/snapshot.py
@@ -31,7 +31,7 @@ from devqubit_braket.utils import (
     obj_to_dict,
 )
 from devqubit_engine.uec.device import DeviceSnapshot
-from devqubit_engine.utils.time_utils import utc_now_iso
+from devqubit_engine.utils.common import utc_now_iso
 
 
 if TYPE_CHECKING:
