+++ packages/devqubit-cirq/src/devqubit_cirq/adapter.py
@@ -63,8 +63,8 @@ from devqubit_engine.uec.types import (
     ProgramRole,
     TranspilationMode,
 )
+from devqubit_engine.utils.common import utc_now_iso
 from devqubit_engine.utils.serialization import to_jsonable
-from devqubit_engine.utils.time_utils import utc_now_iso
 
 
 logger = logging.getLogger(__name__)
