+++ packages/devqubit-pennylane/src/devqubit_pennylane/snapshot.py
@@ -24,8 +24,8 @@ import logging
 from typing import TYPE_CHECKING, Any
 
 from devqubit_engine.uec.device import DeviceSnapshot, FrontendConfig
+from devqubit_engine.utils.common import utc_now_iso
 from devqubit_engine.utils.serialization import to_jsonable
-from devqubit_engine.utils.time_utils import utc_now_iso
 from devqubit_pennylane.utils import (
     collect_sdk_versions,
     extract_shots_info,
