+++ packages/devqubit-qiskit-runtime/src/devqubit_qiskit_runtime/adapter.py
@@ -79,8 +79,8 @@ from devqubit_engine.uec.types import (
     ProgramRole,
     TranspilationMode,
 )
+from devqubit_engine.utils.common import utc_now_iso
 from devqubit_engine.utils.serialization import to_jsonable
-from devqubit_engine.utils.time_utils import utc_now_iso
 from devqubit_qiskit.serialization import QiskitCircuitSerializer
 from devqubit_qiskit_runtime.circuits import (
     circuits_to_text,
