+++ packages/devqubit-qiskit-runtime/src/devqubit_qiskit_runtime/envelope.py
@@ -16,7 +16,7 @@ from typing import Any
 from devqubit_engine.core.run import Run
 from devqubit_engine.uec.envelope import ExecutionEnvelope
 from devqubit_engine.uec.result import ResultSnapshot
-from devqubit_engine.utils.time_utils import utc_now_iso
+from devqubit_engine.utils.common import utc_now_iso
 from devqubit_qiskit_runtime.utils import get_backend_name, get_backend_obj
