+++ packages/devqubit-qiskit-runtime/src/devqubit_qiskit_runtime/snapshot.py
@@ -24,8 +24,8 @@ import logging
 from typing import TYPE_CHECKING, Any
 
 from devqubit_engine.uec.device import DeviceSnapshot, FrontendConfig
+from devqubit_engine.utils.common import utc_now_iso
 from devqubit_engine.utils.serialization import to_jsonable
-from devqubit_engine.utils.time_utils import utc_now_iso
 from devqubit_qiskit.snapshot import create_device_snapshot as create_backend_snapshot
 from devqubit_qiskit_runtime.utils import (
     collect_sdk_versions,
