+++ packages/devqubit-qiskit/src/devqubit_qiskit/snapshot.py
@@ -15,8 +15,8 @@ from typing import TYPE_CHECKING, Any
 
 from devqubit_engine.uec.calibration import DeviceCalibration
 from devqubit_engine.uec.device import DeviceSnapshot
+from devqubit_engine.utils.common import utc_now_iso
 from devqubit_engine.utils.serialization import to_jsonable
-from devqubit_engine.utils.time_utils import utc_now_iso
 from devqubit_qiskit.calibration import extract_calibration_from_properties
 from devqubit_qiskit.utils import get_backend_name, qiskit_version
